+++ Source/MTAFRAT/plugins/iPtorrents/iPtorrents.vb
@@ -18,7 +18,7 @@ Class IPtorrentsPlugin : Inherits DynamicPlugin
     '
     ' VIP / Paid account registration URL: https://www.iptorrents.com/card-pt.php
 
-    ReadOnly headless As Boolean = False
+    ReadOnly headless As Boolean = True
     ReadOnly additionalArgs As String() = Array.Empty(Of String)()
 
     Overloads Async Function RunAsync() As Task(Of RegistrationFlags)
@@ -29,8 +29,11 @@ Class IPtorrentsPlugin : Inherits DynamicPlugin
                 Using service As ChromeDriverService = Nothing,
                       driver As ChromeDriver = CreateChromeDriver(Me, service, headless, additionalArgs)
 
+                    Const triggerRegistration As String = "Payment"
                     Try
-                        regFlags = regFlags Or Me.CustomRegistrationCheck(driver)
+                        regFlags = regFlags Or
+                                   PluginSupport.DefaultRegistrationFormCheckProcedure(Me, driver, triggerRegistration,
+                                                                                                   isOpenTrigger:=False)
 
                     Catch ex As Exception
                         PluginSupport.LogMessageFormat(Me, "StatusMsg_ExceptionFormat", ex.Message)
@@ -45,34 +48,6 @@ Class IPtorrentsPlugin : Inherits DynamicPlugin
 
                 Return regFlags
             End Function)
-    End Function
-
-    Private Function CustomRegistrationCheck(driver As ChromeDriver) As RegistrationFlags
-
-        PluginSupport.LogMessageFormat(Me, "StatusMsg_ConnectingFormat", Me.Name)
-        PluginSupport.LogMessage(Me, $"➜ {Me.UrlRegistration}")
-        PluginSupport.NavigateTo(driver, Me.UrlRegistration)
-
-        PluginSupport.LogMessage(Me, "StatusMsg_CloudflareTrialWait")
-        PluginSupport.WaitForPageReady(driver)
-        Thread.Sleep(5000)
-        Dim selector As By = By.CssSelector(".loginContent")
-        PluginSupport.WaitForElement(driver, selector)
-        PluginSupport.LogMessage(Me, "StatusMsg_CloudflareTrialCompleted")
-        PluginSupport.WaitForPageReady(driver)
-        PluginSupport.LogMessage(Me, "StatusMsg_RegisterPageLoaded")
-
-        Dim pageSource As String = driver.PageSource
-        PluginSupport.LogMessage(Me, "StatusMsg_AnalyzingPageContent")
-        If pageSource.Contains("Credit Card", StringComparison.InvariantCultureIgnoreCase) OrElse
-                           pageSource.Contains("pic/signup.png", StringComparison.InvariantCultureIgnoreCase) Then
-            PluginSupport.LogMessage(Me, "StatusMsg_DetectedRegClosed")
-            Return RegistrationFlags.RegistrationClosed
-        Else
-            PluginSupport.LogMessage(Me, "StatusMsg_DetectedRegOpen")
-            PluginSupport.NotifyMessageFormat("😄🎉🎉🎉", MessageBoxIcon.Information, "StatusMsg_MsgboxRegOpenFormat", Me.Name)
-            Return RegistrationFlags.RegistrationOpen
-        End If
     End Function
 
 End Class